refactor: extract GameSearchService and GameProfileService (steps 5-6)
Step 5: move searchGamesAutocomplete and searchGames into
GameSearchService.ts and migrate all callers off Game.search*.

Step 6: move profile/social read methods into GameProfileService.ts:
getGameReleasesDetailed, getGameAssociations, getNowPlayingMembers,
getGameCompletions, getGameCollectionOwners, getGameProfile,
getAllCompanies, getCompanyById, plus the shared getGameRelations
cache and fetchAllPages helper. Game.ts metadata getters now call
GameProfileService.getGameRelations; remaining callers migrated.

Game.ts is now 469 lines (core CRUD, index, image, field updates,
and metadata getters only). tsc, lint, and all 59 tests pass.
