+++ src/events/MessageReactionAdd.command.ts
@@ -45,6 +45,7 @@ import { logError } from "../utilities/LogUtils.js";
 import { toUnixTimestamp } from "../functions/DateFormatUtils.js";
 import GamePlatformRegionService from "../classes/GamePlatformRegionService.js";
 import { saveFullGameMetadata } from "../functions/GameIgdbSync.js";
+import GameSearchService from "../classes/GameSearchService.js";
 
 const PUSH_PIN_EMOJI = "📌";
 const PLUS_EMOJI = "➕";
@@ -297,7 +298,7 @@ export class MessageReactionAdd {
     session.completionType = value as CompletionType;
     session.promptMessageId = interaction.message?.id ?? session.promptMessageId;
     session.promptChannelId = interaction.channelId ?? session.promptChannelId;
-    const matches = await Game.searchGames(session.query);
+    const matches = await GameSearchService.searchGames(session.query);
     if (!matches.length) {
       await this.promptIgdbImport(interaction, session);
       return;
